requirements.txt
asgiref==3.2.3
dataclasses==0.6
Django==2.2.8
django-crispy-forms==1.8.1
django-localflavor==2.2
djongo==1.2.38
Pillow==6.2.1
pymongo==3.10.0
python-stdnum==1.12
pytz==2019.3
sqlparse==0.2.4
